Setup done for Doubly linked list
