+++ src/data_structures/linked_lists/singly_linked_list/singly_linked_list.py
@@ -14,15 +14,3 @@ class SinglyLinkedListNode:
     def __init__(self, data: Any):
         self.data = data
         self.next: Optional[SinglyLinkedListNode] = None
-
-
-class MultipleElementsHandler:
-    def __init__(self, index, element):
-        self.index = index
-        self.element = element
-
-
-def sort_list_multiple_elements_handlers(
-    multiple_elements_handler_list: list[MultipleElementsHandler],
-) -> list[MultipleElementsHandler]:
-    return sorted(multiple_elements_handler_list, key=lambda k: k.index)
