+++ src/data_structures/linked_lists/singly_linked_list/singly_linked_list_operations.py
@@ -1,9 +1,9 @@
-from .singly_linked_list import (
-    SinglyLinkedList,
-    SinglyLinkedListNode,
+from .singly_linked_list import SinglyLinkedList, SinglyLinkedListNode
+from ..linked_list_utilities import (
     MultipleElementsHandler,
     sort_list_multiple_elements_handlers,
 )
+
 from typing import Any, Optional
