+++ tests/test_linked_lists/test_singly_linked_list.py
@@ -6,6 +6,8 @@ import pytest
 from src.data_structures.linked_lists.singly_linked_list.singly_linked_list import (
     SinglyLinkedList,
     SinglyLinkedListNode,
+)
+from src.data_structures.linked_lists.linked_list_utilities import (
     MultipleElementsHandler,
     sort_list_multiple_elements_handlers,
 )
